Use saved alignments when rebuilding in the test app
* refactor ILLWeightTuner interface to IParameterTuner
* save new weights after training a Thot model
* fix bugs in TextFileText and TextFileTextAlignmentCollection
* factor ISegmentAligner. GetBestAlignment() to accept a hint matrix
* revert back to using Simplex for weight tuning
* no longer dispose referenced engines in HybridTranslationEngine

+++ src/SIL.Machine.Translation.Thot/Thot.cs
@@ -160,7 +160,7 @@ namespace SIL.Machine.Translation.Thot
 
 		[DllImport("thot", CallingConvention = CallingConvention.Cdecl)]
 		public static extern void llWeightUpdater_updateClosedCorpus(IntPtr llWeightUpdaterHandle, IntPtr[] references, IntPtr nblists, IntPtr scoreComps, uint[] nblistLens,
-			double[] weights, uint numSents, uint numWeights);
+			float[] weights, uint numSents, uint numWeights);
 
 		[DllImport("thot", CallingConvention = CallingConvention.Cdecl)]
 		public static extern void llWeightUpdater_close(IntPtr llWeightUpdaterHandle);
